I just changed the about and home a bit
did some css3 on the home page - I’m not done with it. I mainly got the
things I wanted done, but I don’t have any of them in place… it’s like
1/4th done :/

I added a signup page with html and css and tried to add it to the
router, but i can’t get it to show - so theres also that. Delete that
if you want. I haven’t started on it.